+++ Movies/app/build.gradle
@@ -26,6 +26,7 @@ repositories {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile 'com.android.support:appcompat-v7:23.1.1'
     compile 'com.squareup.picasso:picasso:2.5.2'
     compile 'net.simonvt.schematic:schematic:0.6.4'
     compile 'net.simonvt.schematic:schematic-compiler:0.6.4'
