+++ Movies/app/src/main/java/com/example/nilanjandaw/movies/Constants.java
@@ -1,0 +1,14 @@
+package com.example.nilanjandaw.movies;
+
+/**
+ * Created by Nilanjan Daw on 02/04/2016.
+ * Project: Movies
+ */
+public class Constants {
+
+    public static final String movieURL = "http://api.themoviedb.org/3/discover/movie?";
+    public static String sortBy = "popularity.desc";
+    public static final String apiKey = "8378a35f9b9693560120205b5cb6dab5";
+    public static final String BASE_URL = "http://api.themoviedb.org/3/";
+
+}
