+++ Movies/app/src/main/java/com/example/nilanjandaw/movies/TrailerAdapter.java
@@ -44,7 +44,7 @@ public class TrailerAdapter extends BaseAdapter {
 
     @Override
     public long getItemId(int position) {
-        return 0;
+        return position;
     }
