+++ Movies/app/src/main/res/layout-xlarge/activity_main.xml
@@ -1,0 +1,28 @@
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:orientation="horizontal"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:layout_weight="1"
+    android:baselineAligned="false">
+
+    <fragment
+        android:id="@+id/fragment_list"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_vertical"
+        android:layout_weight="2"
+        android:name="com.example.nilanjandaw.movies.MovieListFragment"
+        tools:layout="@layout/fragment_movie_list"
+        />
+
+    <FrameLayout
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        android:layout_weight="4"
+        android:layout_gravity="end|top"
+        android:id="@+id/fragment_holder"/>
+
+</LinearLayout>
+
