+++ Movies/app/src/main/res/layout/fragment_movie_list.xml
@@ -1,14 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context="com.example.nilanjandaw.movies.MovieListFragment">
 
     <GridView
         android:id="@+id/grid_movie"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="center_vertical"
+        android:layout_gravity="left|top"
         android:numColumns="auto_fit"
         android:stretchMode="columnWidth" />
 </FrameLayout>
