+++ Movies/app/src/main/res/layout/trailer_layout.xml
@@ -3,6 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center_horizontal"
+    android:layout_marginStart="10dp"
     android:orientation="vertical">
 
     <ImageView
@@ -10,6 +11,7 @@
         android:layout_width="100dp"
         android:layout_height="100dp"
         android:contentDescription="@string/movie_poster"
+        android:layout_marginStart="10dp"
         android:scaleType="centerCrop"
         android:src="@android:drawable/ic_media_play" />
 
@@ -18,6 +20,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
+        android:layout_marginStart="10dp"
         android:gravity="center"
         android:text="Trailer"
         android:textAlignment="center" />
